pldm: set multi-part component flags
Summary:
When doing a firmware update over PLDM, DSP0267 specifies a state
machine for update.  When sending a multi-part image, we are currently
violating the state machine by sending "TFLAG_START_END" on all parts.

Adjust the PLDM code to determine the multi-part count and use the
appropriate flag (START, MID, END, START_END) based on where we are
in the sequence, so as to not violate the state machine.

Test Plan: Terminus team to test in QEMU.

Reviewed By: lpradeepfb

Differential Revision: D49924126

fbshipit-source-id: e11b0258c8838e1936fd2da3674352fa948ec07b