ncsi-util: handle multi-part transfers
Summary:
When we do a multi-part PLDM transfer to a NIC, we are required to
transfer all parts before proceeding to activation.  The previous
code was transferring only a single part and then skipping to
activation.  Adjust the flow so that we track the number of parts
being transferred and expect an APPLY for each of them before proceeding
to activate.

Test Plan: Terminus team to test in QEMU.

Reviewed By: lpradeepfb

Differential Revision: D49924127

fbshipit-source-id: 935820a64a3df9375e338df702c57f5ca2c782da