Remove jam from Meru main image
Summary:
# Description

Removing jam from meru-image.bb because the main image does not need it since Meta already provides it. Also, show-tech was moved down to keep the installs in alphabetical order.

X-link: https://github.com/facebookexternal/openbmc.arista/pull/322

Test Plan: The build was failing because jam was not meant to be included in the main image. The build now passes without it.

Reviewed By: tao-ren

Differential Revision: D51420374

Pulled By: joancaneus

fbshipit-source-id: 0244d8950d719fddef226346feaf6100e963fed0