rackmon2: Use a shared_mutex to protect threads
Summary:
Instead of expecting the user from doing the right
thing, correctly project the threads pointer instead
of relying only on the reference counting of the
shared pointers.

Use reader/writer locks (shared_mutex) to ensure that
we are allowing some overlap when it comes to to
the most often used (read) functions

Test Plan:
TSAN errors about race to monitorThread_ and
scanThread_ are fixed.

Reviewed By: tao-ren

Differential Revision: D52272301

fbshipit-source-id: 01783e69063ae8dc52c411d02cf584cd1328b817