rackmon2: Dont use 1s polling for tests
Summary:
This was needed when we did not have the ability
to "tick" the threads and the test code had to
rely on a series of sleeps to test the functionality.

Now those tests have been enhanced, we no longer
need to start the monitor with 1s polling interval
and instead let it pick the default 1min.

This fixes a rare issue where the polling thread
ticks on its own along with our test thus causing
unexpected behavior.

Test Plan: All unit tests pass and no delay is observed.

Reviewed By: doranand

Differential Revision: D52349709

fbshipit-source-id: 64a0c75e79d061bd8c1b7f956fd8fe2b5c9ec83a