mfg-tool: add 'hello world' application
Summary:
Since the Redfish support is not complete for the manufacturing support
necessary for the LF-OpenBMC systems, we've been asked to create
shell-based wrappers for EVT to enable manufacturing.  Create a basic
framework of an application that we can add sub-tools to.

This uses meson to build and has "subproject" support, which allows
it to be built outside of Yocto for quick developer.  Yocto recipe will
be added with another commit.

Test Plan:
Built on an x86 system and tested.
```
$ builddir/mfg-tool hello
<6> Hello World!
[
    "Hello World!"
]
$ builddir/mfg-tool hello 2&> /dev/null
[
    "Hello World!"
]
```

Reviewed By: amithash

Differential Revision: D52750377

fbshipit-source-id: 522ca44c1015622e7ec4d9c705fde5a55f1d6c14