mfgtool: add power-state command
Summary: Add a command to get the power state of the slots.

Test Plan:
Tested on QEMU.  Don't have access to Yv4 hardware presently.

It is quite possible these values are completely bogus and might not
even be right on the real hardware, but at least the right structure
comes back for now.
```
{
    "0": {
        "standby": "off"
    },
    "1": {
        "runtime": "on",
        "standby": "off"
    },
    "2": {
        "runtime": "on",
        "standby": "off"
    },
    "3": {
        "runtime": "on",
        "standby": "off"
    },
    "4": {
        "runtime": "on",
        "standby": "off"
    },
    "5": {
        "runtime": "on",
        "standby": "off"
    },
    "6": {
        "runtime": "on",
        "standby": "off"
    },
    "7": {
        "runtime": "on",
        "standby": "off"
    },
    "8": {
        "runtime": "on",
        "standby": "off"
    }
}
```

Reviewed By: amithash

Differential Revision: D52882005

fbshipit-source-id: 14d067a4bea1560260a0c13d100ca01776baffb1