mfgtool: power-state: add more granularity to state
Summary:
While power state transitions are in progress, the
state property on the Host and Chassis can be other than "off" and "on"
for longer than expected.  Add more complete state mapping to advertise
this correctly.

Test Plan:
In QEMU, the "wait-until-mctp-connection-done" doesn't complete, so
the chassis power-on does not complete (3).  If you kill the process, then
the transition will complete (1).  If you then issue a host power-off,
the chassis stays powered on correctly (2).

```
    "1": {
        "runtime": "on",
        "standby": "on"
    },
    "2": {
        "runtime": "off",
        "standby": "on"
    },
    "3": {
        "runtime": "transition-on",
        "standby": "transition-on"
    },
```

Reviewed By: wangx6f

Differential Revision: D53436083

fbshipit-source-id: 186f518d0d3b63145a0570e3657d14510836c94e