Elbert: Fix power on/setup i2c logic in CPU mode
Summary:
# Description

Fix the power on and setup i2c scripts to exit rather then return
when in CPU mode.

# Motivation:

When the system is in CPU mode, there's a potential for i2c devices to
conflict with accesses from the CPU.

This fix was actually done in the following PR:
https://github.com/facebookexternal/openbmc.arista/pull/106

But the code should have been exiting rather than returning. It's not
clear if something changed that caused this code to behave differently
or if this code always behaved this way.

X-link: https://github.com/facebookexternal/openbmc.arista/pull/369

Test Plan:
1. Booted the BMC and verified that there are appropriate messages in
the boot log indicating that device setup is skipped when not in BMC mode.

Specifically, verified the absence of "Instantiated device" logs.

2. Verified when in BMC mode the BMC boots up normally and i2c devices
are setup.

Here's a specific example of an "Instantiated device" log seen when in BMC mode:
```
2024 Jun 17 10:06:29 bmc-oob. kern.info elbert-929a0a120d0-dirty: kernel: i2c i2c-11: new_device: Instantiated device max6658 at 0x4c
```

Reviewed By: alandau

Differential Revision: D59654770

Pulled By: joancaneus

fbshipit-source-id: abdc5e2f38bba4855ff13d678f5fa5b5c95f20d5