qemu: Fix flash size mismatch error
Summary:
Qemu sometimes fails with
```
qemu-system-aarch64: device requires 67108864 bytes, block backend provides 134217728 bytes
```

This started happening after the upgrade to qemu 8.2.1, and is unrelated
to the elbert fix D60263812 (reverting the fix, I'm still seeing this
error with qemu 8.2.1).

Code in our patch `0501-aspeed-Zero-extend-flash-files-to-128MB.patch`
zero-extends the image file given on the command line to the size of the
flash mmap region. This used to work in qemu 7.2.0, but in 8.2.1 for
some unknown to me reason, the mmap region is always 128MB and is not
dependent on the specified flash size. Therefore any flash size except
128MB causes the error.

This diff changes the equality test to succeed if the flash size
is <=128MB, which works on all platforms I tested:
wedge100, wedge400, elbert, fuji, minipack, yamp.

Test Plan:
For all of the above platforms:
```
build/qemu-system-aarch64 -machine wedge100-bmc -nographic -drive file=/home/alandau/openbmc/build-wedge100/tmp/deploy/images/wedge100/flash-wedge100,format=raw,if=mtd,snapshot=on
```

And check that `/proc/mtd` and `dmesg|grep spi` list the correct flash size.

Reviewed By: williamspatrick

Differential Revision: D61751440

fbshipit-source-id: 9b2bb1754cd473d1dbe7f9a023c2e912262f5e9a