libcper: copy recipe from upstream
Summary:
Copy the latest version of the cper recipe from the
openbmc/openbmc repository.  This includes the CPER patches
and removes the dependency on base64.

Disable ptest support in libcper for now because it is failing
to generate a proper ptest runner.

Test Plan: Builds pass.

Reviewed By: doranand

Differential Revision:
D61811416

Privacy Context Container: L1188897

fbshipit-source-id: 6ba078fec5439347459bf6d5008ac3d434e2fd77