rackmon2: Add support for ORv3 PSU/BBU on AALC Rack
Summary:
When adding support for HPR, we narrowed the allowed
address range of ORv3 PSU/BBU to be exactly what we
expect on ORv3 racks.

This unfortunately removed support for ORv3 PSUs/BBUs
which could be connected to Port 3.

Add ranges to support them.

Test Plan:
Copy new configs to HPR rack wedge400-bmc and replace ones in
/etc/rackmon.d. Restart rackmond.

We should be seeing the new PSUs/BBUs now.

Reviewed By: wangx6f

Differential Revision: D62591372

fbshipit-source-id: 31ffd6a43a28fbf098c52f8429258ee69715645a