GTT/GTI: fix detecting unknown GPU config
Summary:
# Description
The issue was found on H100 with FW bundle v.1.4.0. In case of failure to retrieve a byte from the GPU Baseboard FRU EEPROM. So, we add a retry mechanism as a workaround.

X-link: https://github.com/facebookexternal/openbmc.quanta/pull/4442

Test Plan: Build and test pass on GTT/GTI H100 system

Reviewed By: amithash

Differential Revision: D63319965

fbshipit-source-id: 15a9c5b82232127cbfb2c3e8b211611429f13610