rackmon2: Remove support for baudrate negotiation
Summary:
Completely remove all code and config associated with
baudrate negotiation, any configuration linking to it
and remove config keys "default_baudrate" and
"preferred_baudrate" in favor of a single "baudrate".

While the intention of the feature was in the correct
place, it poses more stability issues and additional
testing with no real-world measurable benefit. Worse,
if rackmon/bmc crashes when the baudrate is negotiated
up, when rackmon comes back up, it would not be able
to discover the unit in the "default" baudrate.

Considering its lack of use both current and future,
it has minimal value to carry forward and add to the
maintenance cost.

Test Plan:
All unit tests pass
Copy to wedge400 and ensure it is functional.

Reviewed By: doranand

Differential Revision:
D63329816

Privacy Context Container: L1188897

fbshipit-source-id: 002207822a7b6db75ff2e2213e5f5771356c8574