Yamp: DC PSU Support
Summary:
# Description

Modify the setup_i2c.sh script to check the model of the PSU before adding an i2c device for it.

# Motivation

When DC PSU's are present, writing to the WRITE_PROTECT register or adding the i2c device causes the PSU to get shut off. As a result, OpenBmc is not usable for Yamp when DC PSU's are present.

X-link: https://github.com/facebookexternal/openbmc.arista/pull/374

Test Plan:
1. On a system with DC PSU's, switched to BMC mode and verified the system did not get powered down and the system was usable.

2. On a system without DC PSU's, verified that the PSU i2c devices were present:
```
/sys/bus/i2c/devices/5-0058
/sys/bus/i2c/devices/8-0058
```

Reviewed By: alandau

Differential Revision: D63722920

Pulled By: joancaneus

fbshipit-source-id: 25d5ca43c3f635c69b1f4824aa4e7a8fc05f14ed