Adding pre_action entry similar to the existing post_action
Summary:
UFW already support a json entry called post_action, but we
don't have any pre_action. The pre_action is needed in case we need to
execute a command before running the actual upgrade step. We never had a
need for pre-upgrade but with wedge400, we just came across the need to
run a command before the actual upgrade steps.

ufw is part of the flasher fpkg and the fix_COME.sh script will also be
part of the binary and manifest json packages.

Test Plan:
run ufw on a wedge400 host and verify that it does the
pre_upgrade command. P1635895537

The modified json looks like this P1635903774 (see pre_action). we will
make a diff to land this json as well in openbmc.

Reviewed By: alandau

Differential Revision: D64003563

fbshipit-source-id: 1bfda4b1b23236cb3748edbb253a0c3538870986