meta-facebook: yosemite4: Update GPIO monitor json (#3214)
Summary:
# Description:
This commit updates the GPIO monitor configuration JSON to align with the changes proposed in Gerrit 75557.
It ensures that the necessary modifications will be properlyrecognized and integrated into the system.
Gerrit link: https://gerrit.openbmc.org/c/openbmc/openbmc/+/75557

# Motivation:
This update addresses the issue where PLDM devices disappear during a 12V power cycle. For the resolution of this problem, both Gerrit 75557 and this PR need to be merged together, as the changes are interdependent for the successful initialization of the PLDM devices.

Tested:
Completed 10 rounds:
1. Successfully verified PLDM devices appear after performing 12V power cycles sequentially on all 8 slots - PASS
2. Build code - PASS

# Description

Please include a summary of the change and which issue is fixed.

# Motivation

Please include an explanation of why you these changes are necessary

Pull Request resolved: https://github.com/facebookexternal/openbmc.wiwynn/pull/3214

Test Plan: Please describe the tests that you ran to verify your changes. Provide instructions so we can reproduce.

Reviewed By: williamspatrick

Differential Revision: D65532351

fbshipit-source-id: da2c1e3c4857f77e37a7428e9a6437ea1381bccc