The new AURA Clock generator clock / PLL fix
Summary:
[FUJI] The new AURA Clock generator clock / PLL fix
As per the advice from HW team, we will still honor the previous version
of the fix, and not apply the fix, if the previous fix is already
applied.
What the new fix does is essentially similar to what the previous fix
does, but in a different way.
This new fix includes the logic to honor the old fix and not apply the
new fix if the old fix is already applied.

Test Plan:
After reset (i2c) commands, the new fix works : P1680204231
After reset (i2c) commands, if we run the legacy fix first, then the new
fix will not kick in : P1680204231

Reviewed By: joancaneus

Differential Revision:
D66034889

Privacy Context Container: L1125642

fbshipit-source-id: 0aa0f572e7036ae492da7878f3debd481d7561fc