flashy: remediation: disable kernel hung_task_panic
Summary:
Disable the "hung_task_panic" as a common remediation.

We have seen on some Yv4 systems this kernel panic.  The default
behavior per the kernel documentation is disabled, but there is
a lesser documented CONFIG_BOOTPARAM_HUNG_TASK_PANIC which is the
real default (and which somehow is enabled presently).  By setting
to disabled we can reduce this type of kernel panic during flash
updates.

Test Plan: Go tests.

Reviewed By: doranand

Differential Revision: D66163794

fbshipit-source-id: 37b9363800cc24a5c69f13c61dcef720dedd25ac