weutil: Detect EEPROM v6
Summary:
Arista brought to my attention that when adding v6 support, I
forgot to update one place that does version detection. By lucky
coincidence, when no version is matched (the case we're hitting) it
returns an error message as a string, and later compares it to a
predefined `ARISTA_PREFDL` string, and doing the right thing when they
are not equal.

In other words, this diff is not really necessary, but protects against
future changes that might make the accidental success a failure.

This, along with a lot more code, should be deleted when we get rid of
the old parser.

Test Plan: Eyeballing, CI

Reviewed By: somasun

Differential Revision: D67311997

fbshipit-source-id: dc85282b389608ee599e7551c7e479a247747af7