mfg-tool: log-registry: command search registry for events
Summary:
Add a command that allows users to search the Redfish
Message Registries for a corresponding OpenBMC event.  Redfish
events are what are documented in tooling specs but OpenBMC
events are what are used by the `log-create` utility.  This utility
is useful for discovering the OpenBMC events that might result
in a particular Redfish event (for the purposes of error injection).

Test Plan:
Built on development system and ran:
```
$ mfg-tool log-registry "OpenBMC_StateLeakDetector"
<6> Loading registries from .../builddir/subprojects/phosphor-dbus-interfaces/registry
[
    "xyz.openbmc_project.State.Leak.Detector.LeakDetectedWarning",
    "xyz.openbmc_project.State.Leak.Detector.LeakDetectedNormal",
    "xyz.openbmc_project.State.Leak.Detector.LeakDetectedCritical"
]
```

Reviewed By: amithash

Differential Revision:
D68308527

Privacy Context Container: L1188897

fbshipit-source-id: d399207b6689430a738392905da305bc882f48de